Refactor Box
