+++ src/js/components/Box/index.d.ts
@@ -23,7 +23,6 @@ export interface BoxProps {
   responsive?: boolean;
   round?: boolean | "xsmall" | "small" | "medium" | "large" | "xlarge" | "full" | string | {corner?: "top" | "left" | "bottom" | "right" | "top-left" | "top-right" | "bottom-left" | "bottom-right",size?: "xsmall" | "small" | "medium" | "large" | "xlarge" | string};
   tag?: string;
-  as?: string;
   width?: "xsmall" | "small" | "medium" | "large" | "xlarge" | string;
   wrap?: boolean;
 }
